pyproject.toml
[project]
name = "craft-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "discord-py>=2.5.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]
